KeyValuePassMemTable::fileno: needs strict uint64
In CSPPMemTable, value ref to wal need not full uint64, uint32 is
sufficient, because fileno of different wal files in a MemTable can
not span UINT32_MAX, store uint64(as text string) in meta data is
is suffient.

+++ include/rocksdb/write_batch.h
@@ -48,7 +48,7 @@ struct KeyValuePassMemTable {
   Slice    value;
   uint64_t val_pos;
   uint32_t key_len;
-  uint32_t fileno;
+  uint64_t fileno;
   const ReadonlyFileMmap* wal_file;
 };
